+++ ios/tiktok_events_sdk/Sources/tiktok_events_sdk/FlushHandler.swift
@@ -1,0 +1,10 @@
+import Flutter
+import TikTokBusinessSDK
+import Foundation
+
+struct FlushHandler {
+    static func handle(call: FlutterMethodCall, result: @escaping FlutterResult) {
+        TikTokBusiness.explicitlyFlush()
+        result("TikTok flush triggered")
+    }
+}