+++ ios/tiktok_events_sdk/Sources/tiktok_events_sdk/GetIdfaHandler.swift
@@ -1,0 +1,9 @@
+import Flutter
+import TikTokBusinessSDK
+import Foundation
+
+struct GetIdfaHandler {
+    static func handle(call: FlutterMethodCall, result: @escaping FlutterResult) {
+        result(TikTokBusiness.idfa())
+    }
+}