+++ ios/tiktok_events_sdk/Sources/tiktok_events_sdk/IsTrackingEnabledHandler.swift
@@ -1,0 +1,9 @@
+import Flutter
+import TikTokBusinessSDK
+import Foundation
+
+struct IsTrackingEnabledHandler {
+    static func handle(call: FlutterMethodCall, result: @escaping FlutterResult) {
+        result(TikTokBusiness.isTrackingEnabled())
+    }
+}